Simplify msglen to just use ints

+++ main.c
@@ -119,7 +119,7 @@ do_agent_loop(int sockfd)
             }
             else {
                 int len = read(fd, buf, sizeof(buf));
-                if (len >= 4 && len == (int)msglen(buf)) {
+                if (len >= 4 && len == msglen(buf)) {
                     sendbuf[fd] = agent_query(buf);
                     FD_SET(fd, &write_set);
                 }
@@ -132,7 +132,7 @@ do_agent_loop(int sockfd)
         FD_FOREACH(fd, &do_write_set) {
             void *reply = sendbuf[fd] ?: &reply_error;
             int len = write(fd, reply, msglen(reply));
-            if (len == (int)msglen(reply))
+            if (len == msglen(reply))
                 FD_SET(fd, &read_set);
             else
                 close(fd);
